requirements.txt
python-telegram-bot==20.7
yt-dlp==2023.11.16
requests==2.31.0
beautifulsoup4==4.12.2
playwright==1.40.0
aiohttp==3.9.1
python-dotenv==1.0.0
Pillow==10.1.0
